chore: add status code for request_total metrics
